modal popup

+++ ui/main.js
@@ -1,44 +1,67 @@
+/**
+ * Callback for click on service list item.
+ * @param {string}   service - Service name.
+ * @param {DOMEvent} event   - Click event.
+ */
 function onServiceClick(service, event) {
-    for(var i = 0; i < event.target.classList.length; i++)
+    for(var i = 0; i < event.target.classList.length; i++) {
         if(event.target.classList[i] == 'delete') return;
-    alert(service);
+    }
+    modal(service);
 }
 
+/**
+ * Callback for clickc on service delete button.
+ * @param {string}   service - Service name.
+ * @param {DOMEvent} event   - Click event.
+ */
 function onDeleteClick(service, event) {
-    if(confirm("Do you really want to delete " + service + "?"))
-        manager.delete(service);
+    modal('Do you really want to delete ' + service + '?', function(res) {
+        alert(res.button);
+        if(res.button === 'YES') manager.delete(service);
+    }, ['YES', 'NO']);
 }
 
-function modal(text, button, result) {
-
-}
-
+/**
+ * Add service to list.
+ * @param {string} service - Service name.
+ */
 this.addService = function addService(service) {
     var list = document.getElementById('servicelist');
-    var li = document.createElement('li');
-    var cont = document.createElement('div');
-    var a = document.createElement('a');
-    li.classList.add('service');
-    li.classList.add('box');
+    var li = createNode('li', null, ['service', 'box']);
+    var cont = createNode('div', null, ['box'], service);
+    var a = createNode('a', null, ['delete']);
     li.onclick = onServiceClick.bind(null, service);
-    a.classList.add('delete');
     a.onclick = onDeleteClick.bind(null, service);
-    cont.classList.add('box');
     li.appendChild(cont);
     li.appendChild(a);
     list.appendChild(li);
-    cont.innerHTML = service;
 }
 
-this.clearList = function clearList() {
+/**
+ * Clear service list.
+ */
+function clearList() {
     var list = document.getElementById('servicelist')
     while (list.firstChild) {
         list.removeChild(list.firstChild);
     }
 }
 
-this.setMaster = function setMaster() {
-    modal('Set the master password', 'SET', function(res){
-        manager.setMaster(res);
-    });
+/**
+ * Set master password.
+ */
+function setMaster() {
+    modal('Set the master password', function(res){
+        manager.setMaster(res.password);
+    }, ['SET', 'CANCEL'], [{
+            type: 'password',
+            name: 'password',
+            hint: 'Password',
+        }]
+    );
 }
+
+// Export functions
+this.clearList = clearList;
+this.setMaster = setMaster;
